refactor: Extract common parts
